Consolidate TokenList Actions
