+++ Authenticator/Source/TokenList.swift
@@ -58,3 +58,19 @@ struct TokenList {
         return Array(periods).sort()
     }
 }
+
+extension TokenList {
+    enum Action {
+        case BeginAddToken
+        case BeginEditPersistentToken(PersistentToken)
+
+        case UpdatePersistentToken(PersistentToken)
+        case CopyPassword(String)
+
+        case MoveToken(fromIndex: Int, toIndex: Int)
+        case DeleteTokenAtIndex(Int)
+
+        // TODO: remove this action and have the component auto-update the view model on time change
+        case UpdateViewModel
+    }
+}
